AMOYRM1B/requirements.txt
python-telegram-bot==22.1
httpx==0.28.1  # (Only if your bot uses HTTP requests)
python-dotenv==1.0.0  # (Optional, for .env files)